feat(buildog): outsourced fetching from page.tsx
